refactor: split off transformers for vitest and jest into separate namespace for clarity
